requirements.txt
langchain
PyPDF2
python-dotenv
streamlit
openai
faiss-cpu
altair
tiktoken
tokenizers
torch
INSTRUCTOR
huggingface-hub
cohere
InstructorEmbedding
sentence-transformers
streamlit
langchain_community
langchain_text_splitters
langchain_core


